made some changes...still a lot to work on

+++ main.py
@@ -10,7 +10,7 @@ app = Flask("Fitness Something Something")
 CORS(app)
 
 
-@app.route('/api/sign-up', methods=["POST"])
+@app.route('/sign-up', methods=["POST"])
 def signup():
     try:
         body = request.json
@@ -27,7 +27,7 @@ def signup():
         return error, 500
 
 
-@app.route('/api/sign-in', methods=["POST"])
+@app.route('/sign-in', methods=["POST"])
 def sign_in():
     body = request.json
     conn = create_connection(database)
@@ -70,13 +70,27 @@ def sign_in():
     return '', 204
 
 
-@app.route('/api/sign-in/user-data', methods=["GET"])
+@app.route('/user-data', methods=["GET"])
 def user_data():
     try:
         global current_user
+        response = {
+            "user": []
+        }
         conn = create_connection(database)
         data = get_user_data(conn, current_user)
-        return data, 204
+        print(data)
+        for row in data:
+            element = {
+                "user_name": row[1],
+                "user_age": row[2],
+                "user_gender": row[3],
+                "user_weight": row[4],
+                "user_activity_level": row[5],
+                "user_height": row[6]
+            }
+        response["user"].append(element)
+        return response, 200
     except Exception as e:
         error = {
             'error': f"Failed to get user data. Message:  {e}"
@@ -85,10 +99,15 @@ def user_data():
         return error, 500
 
 
-@app.route('/api/delete', methods=["POST"])
+@app.route('/delete', methods=["POST"])
 def delete():
-    print(current_user)
     conn = create_connection(database)
+    if current_user is None:
+        error = {
+            "error": f"--User does not exist."
+        }
+        conn.close()
+        return error, 404
     try:
         delete_user(conn, current_user)
         return '', 201
